cloud-init optional

+++ patches/ubuntu/files/after-reboot.sh
@@ -6,8 +6,10 @@ tune2fs -c 0 $(cat /proc/self/mounts | grep " / " | cut -f 1 -d " ")
 
 rm -rf /var/lib/apt/lists
 
-cloud-init clean --logs
-rm -rf /var/lib/cloud/*
+if command -v cloud-init; then
+  cloud-init clean --logs
+  rm -rf /var/lib/cloud/*
+fi
 
 # ensure no ssh keys are present
 rm -f /home/ubuntu/.ssh/authorized_keys /root/.ssh/authorized_keys
